requirements.txt
requests
textblob
python-telegram-bot
